Break dependency between storage package and postgresql/sqlite ones and split storage.go file.
